fix: extend IServer interface
