+++ packages/grpc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@join-com/grpc",
-  "version": "0.3.7",
+  "version": "0.3.8",
   "description": "gRPC library",
   "license": "MIT",
   "scripts": {
